Time: 13 ms (64.10%), Space: 9.6 MB (35.21%) - LeetHub

+++ 576-out-of-boundary-paths/576-out-of-boundary-paths.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     long mod=10e8+7;
-    int recursion(int i,int j,int m,int n,int maxMove,vector <vector <vector <long>>> &dp)
+    int recursion(int i,int j,int m,int n,int maxMove,vector <vector <vector <int>>> &dp)
     {
         if(maxMove<0)
             return 0;
@@ -12,7 +12,7 @@ public:
         return dp[i][j][maxMove]=(recursion(i+1,j,m,n,maxMove-1,dp)%mod+recursion(i,j+1,m,n,maxMove-1,dp)%mod+recursion(i-1,j,m,n,maxMove-1,dp)%mod+recursion(i,j-1,m,n,maxMove-1,dp)%mod)%mod;
     }
     int findPaths(int m, int n, int maxMove, int startRow, int startColumn) {
-        vector <vector <vector<long>>> dp(m+1,vector <vector<long>> (n+1,vector <long>(maxMove+1,-1)));
+        vector <vector <vector<int>>> dp(m+1,vector <vector<int>> (n+1,vector <int>(maxMove+1,-1)));
         return dp[startRow][startColumn][maxMove]=recursion(startRow,startColumn,m,n,maxMove,dp);
     }
 };
